7. Move folders around to make it easier to view
